+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes relevant to players in this project will be documented in th
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). This project uses its own versioning system.
 ## [Unreleased]
+### Added
+- Solo players can join a pack prior to realm assignment
 
 ## [1.4.1] - 2021-05-23
 ### Added