+++ CHANGELOG.md
@@ -3,12 +3,19 @@ All notable changes relevant to players in this project will be documented in th
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). This project uses its own versioning system.
 ## [Unreleased]
+
+## [1.4.2] - 2021-06-03
 ### Added
 - Solo players can join a pack prior to realm assignment
 - The Scribes have been updated with additonal information
 
 ### Changed
 - Op Center is now the default Advisors page
+
+### Fixed
+- Bots will no longer overtrain spies and wizards
+- Status page now correctly rounds boats down always
+- Invasions outside of Elite Guard range no longer affect Prestige/School penalties
 
 ## [1.4.1] - 2021-05-23
 ### Added
@@ -1417,7 +1424,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). 
 ### Added
 - This CHANGELOG file.
 
-[Unreleased]: https://github.com/OpenDominion/OpenDominion/compare/1.4.1...HEAD
+[Unreleased]: https://github.com/OpenDominion/OpenDominion/compare/1.4.2...HEAD
+[1.4.2]: https://github.com/OpenDominion/OpenDominion/compare/1.4.1...1.4.2
 [1.4.1]: https://github.com/OpenDominion/OpenDominion/compare/1.4.0...1.4.1
 [1.4.0]: https://github.com/OpenDominion/OpenDominion/compare/1.3.2...1.4.0
 [1.3.2]: https://github.com/OpenDominion/OpenDominion/compare/1.3.1...1.3.2